refactor: refine transform code to be more readable
