+++ packages/playground/src/pages/about.vine.ts
@@ -1,4 +1,4 @@
-import { ref } from 'vue'
+import { ref, Ref } from 'vue'
 //           ^^^^ Deliberately import an extra useless type item here
 //                - For ESLint rules to catch it
 //                - Test if it broke compilation in JS runtime
